Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/worker-importscripts-blocked-expected.txt

PASS result is "importScripts blocked: NetworkError:  A network error occurred."
PASS successfullyParsed is true

TEST COMPLETE
